Move and rename LogToDiagnosticsTests.cs file to Common/InternalLoggerTests.Diagnostics.cs file
