+++ tests/NLog.UnitTests/NLog.UnitTests.netfx35.csproj
@@ -98,6 +98,7 @@
     <Compile Include="AssemblyInfo.cs" />
     <Compile Include="AsyncHelperTests.cs" />
     <Compile Include="Common\InternalLoggerTests.cs" />
+    <Compile Include="Common\InternalLoggerTests.Diagnostics.cs" />
     <Compile Include="Common\LastLogEventListTarget.cs" />
     <Compile Include="Conditions\ConditionEvaluatorTests.cs" />
     <Compile Include="Conditions\ConditionParserTests.cs" />
@@ -185,7 +186,6 @@
     <Compile Include="LogManagerTests.cs" />
     <Compile Include="LogReceiverService\LogReceiverForwardingServiceTests.cs" />
     <Compile Include="LogReceiverService\LogReceiverServiceTests.cs" />
-    <Compile Include="LogToDiagnosticsTests.cs" />
     <Compile Include="NLogTestBase.cs" />
     <Compile Include="NLogTraceListenerTests.cs" />
     <Compile Include="ProcessRunner.cs" />
